Implemented Daemons usage and rake tasks for managing the server process
Added a Manager class that provides a simple interface for managing the server
process. This wraps the Daemons gem and uses it to actually start and stop the
server. The Hosts class contains a list of all service hosts have been defined
using Sanford. This is used to lookup specific service hosts when the rake tasks
are used.