Added in bench task for generating a report, for tracking Sanford's performance
This adds a bench task for testing Sanford's performance. The task generates a
text report of the captured statistics. This way git can keep track of the
historical data of Sanford's performance. This will act as a base and currently
only supports a single request, but will expand to benchmark multiple requests.