Implementation of `Sanford::ServiceHandler` mixin
This commit fleshes out the `ServiceHandler` mixin and adds request logging to
Sanford. The `ServiceHandler` expects the `run!` method to be defined and
provides a `halt` helper method for returning custom response statuses and
halting execution of the service immediately. In addition to this, this commit
includes some organizational changes and adds logging for all requests.