Modified Sanford's system tests to test against a real server running
Previously, Sanford used a fake socket to 'pretend' communicating. The problem
with this is it can hide problems with using a real TCP socket. To handle this,
Sanford's system tests now spin up an instance of a Sanford server in a forked
process and send requests to it. This allows us to test Sanford using an actual
TCP socket without to much overhead.