Updated the `Manager` to use a host class over it's configured name
This makes sure that when a host is specified when using the manager (using
environment variables for example) that it tries to find a host by it's class
before using the configured name. This is to ensure that if a user specifies a
host class when running the rake tasks, that they always get that class.

Closes #8