Support verbose and summary logging
This adds support for verbose and summary logging. This allows for
logging to be switched depending on the environment (or
preference). Verbose logging was the previously implemented
logging and the new summary logging is a condensed single line
with all the available info. There is also a performance boost
with using summary vs. verbose logging, but it's fairly minor.