Make hosts singletons (no longer initialized)
This modifies the `Host` module to create singleton classes when
included. It is no longer initialized by the `Manager`. This
makes it's only responsibility setting and getting a host's
configuration. This change shifts it's initialize logic into a
new utility class `Sanford::Manager::Options`. This class takes
a service host's configuration and optional overrides, and from
them builds options needed by the manager. This allows the
previous behavior of running the same host on multiple ports. The
goal was to improve the readability and make clear the role of
the hosts.