Add performance improvements to configuration and handler lookup
These changes increase the performance of Sanford by preparing
hosts before they are used in the server. This is done in the
new `PreparedHost` class. A host's configuration is merged with
optional overrides and it's configured service handlers are
constantized. This removes overhead from a request's lifecycle,
which translates into increased performance.

Closes #33