Added distribution listing to the standard benchmark report
This adds a listing of the distribution of request times. This
gives a more detailed idea of whats going on. With this
information you can tell if a benchmarking run was a better or
worse run (usually by how many outliers there are). It can
also be used to see if there has been a significant shift in
where most of the requests are landing based on recent changes.