Allow configuring a setup proc for hosts
This will be called before a server is initialized by and
run by the `Manager`. This should be used to load the environment
for the server to run in. This fixes bugs that came up with
loading the environment before using Daemons. Since Daemons is
trying to daemonize the process, it's better to let it do that
first and then load your environment after it's done it.

This also removes defining a `sanford:setup` task. The setup proc
should be used instead of this.