Handle keep-alive connections
This adds handling for keep-alive connections to Sanford. If a
client connects and immediately disconnects, this triggers an
`EndOfStreamError` in the protocol. Sanford will catch this and
based on if the host was expecting keep-alive connections, it will
return a 200 or 400. This is to properly handle when this case
occurs and have default behavior provided for it.