Clean up load order issues
While integrating Sanford, I realized there were some load order
issues that were causing bad values "locked" into `HostData`
instances. This stemmed from 2 issues:

* When the `setup` proc was being called
* The `Manager` class using the `HostData`

In many cases, the `Host` configuration relies on loading the
environment, which is done by the `setup` proc. Thus the `setup`
proc needs to be called before we store the configuration in a
`HostData` instance. Doing this though, causes problems with the
`Manager`. This is because the `setup` proc is now loading the
environment before we have forked (using Daemons), which is why
we added the `setup` proc in a previous commit. The issues is the
coupling of `HostData` to both the `Server` and `Manager`. This
removes the coupling and alters when the `setup` proc is called.