Manually connect the server before running start
This changes Sanford to manually connect it's server (so it starts
queueing up connections) before starting it. This also moves the
creation of the `HostData` to an `on_start` callback. The net
effect is that the server will bind and begin listening before
the `HostData` trys to load the app's environment. This enables
faster start times, which means less dropped connections.