Removed rake helpers
This removes the rake helpers from Sanford and stops using them
for the benchmark tasks. I replaced the benchmark tasks and use
the CLI instead. This is to remove our dependency and usage of
Rake, which the CLI replaces. I also changed the port number used
by the bench server/runner because it had chances to conflict with
the test suite (if the server was running it blew up the test
suite).