Remove daemons
This is to better support features like "hot restarting" and to
provide more options. Daemons locked us in on how it daemonized
processes. In particular, it caused problems with using
`Kernel.exec` to replace the current process with a new command.
This adds a simpler daemonization process and removes the
dependency.