Updated to the latest DatTCP with adjusted API
This updates DatTCP to use it's API that adds more control over
the server. This is to setup allowing building servers from file
descriptors and also "pausing" them. This modifies the interface
for servers and slightly changed the way the manager runs the
server. "Restarting" is currently disabled. This will be
reimplemented as a "hot restart".