Performance tweaks on sockets Sanford interacts with
This adds the TCP_NODELAY option to Sanford's server socket and
corks and uncorks client sockets with TCP_CORK (only on Linux).
Both of these do similar things, and are related to disabling
Nagle's algorithm for improved performance. Nagle's algorithm is
ideal for applications, like Telnet, that send sporadic packets
(as a user types). For HTTP (and Sanford) it isn't ideal, because
all the data is ready to be sent at once, not sporadically.