Added a configuration option for the restart dir
This adds a configuration option for specifying the Sanford
restart directory. Previously, this always used the current
working directory (pwd), but it's convenient to be able to
control this and not rely on that always.

This also centralizes the handling of the client fds. Previously,
they were handled when running the server while all other ENV vars
were handled in the `Config`. This moves them to the `Config`.