Updated the readme and general documentation
This updates the documentation to be relevant with the recent
changes to the gem. In particular, the removal of the rake
tasks and the addition of the CLI to replace them. This also adds
a documentation link to explain the daemonization process and
why it is done like it is.

Closes #55