Remove managing tests for now, issues with forked processes
These tests are not properly handling the processes they spawn
which causes processes to be left running. This can cause lots of
problems when running the test suite multiple times in a row.
For now, I'd rather remove these tests, and re-think how to
properly test this logic. The unit tests should provide decent
coverage, and the bench host can always be started to tests the
full stack.