Add `Sanford` tag to log messages
This modifies Sanford's logging to prepend a `[Sanford]` tag to
all it's log messages. This is for identification purposes of
where the logging is coming from. In addition to this, I
implemented some summary line cleanups from `AndSon`. This moves
the `time` and `status` to the front of the log messages to
help visually read those values.

Closes #67