fail nicely if trying to write the pid file to a dir that is not writeable
Before this would error with a cryptic `Errno::ENOENT` error that went
something like "No such file or directory - /path/to/pid/file.pid".

This changes it to raise a `RuntimeError` that says "PID file dir
`/path/to/pid` is not writeable".

The whole point of this is to help the user out and to better direct
them to fix the problem.

Closes #63.