Don't re-daemonize the process when restarting
Currently, Sanford re-daemonizes it's process when it is
restarted. This isn't ideal because the PID changes, which isn't
the intent with a "hot" restart using exec. This seems to also
cause issues with some monitoring tools losing the process they
should be monitoring. This changes Sanford to explicitly not
run it's daemonize when it's restarted, which should keep the
PID intact.