recursively sanitize response objects
BSON fails to serialize on certain data types, specifically Date
and DateTime.  This sanitizes all response datas to prevent BSON
serialization failures.

Note, this assumes a lot on behalf of the user.  Particularly about
how to convent values that need sanitizing.  Plus you incure the
sanitization overhead on _every_ response, whether sanitation is
needed or not.