test runner now serializes response data
This makes for a more robust test.  It will alert devs to serialization
errors in their tests.

This is useful b/c certain versions of BSON error on certain data
types (like `Date` and `DateTime`).  This means your tests can not
only validate the response data itself, but that it will also
serialize.