force string names when defining services on hosts
This all defined service names to be strings.  This is to avoid
confusing 404 errors when defining as a symbol but calling on the
client with a string.

Note, redding/sanford-protocol similarly forces all requests to
have string names.  This ensures that requests will always be routed
top handlers using string names.