configure template source
This updates the config API to allow configuring a template source.
This is part of the effort to add a render pipeline to Sanford.  To
render a template, it needs to exist in a configured source.  Setting
a template source means this source will be used by default in render
calls.

Note, this involved changing the config proxy to a class instance
implementation.  This instance will track the sources.  I also updated
the configs tests to work accordingly and made them more robust.