sanford runner; remove runner configuration
This removes the concept of a default runner and configuring runners
on hosts.  Instead, there is now only a `SanfordRunner`, which hosts
use to run their handlers, and a `TestRunner` which you use to run
handlers in tests.  The sanford runner is basically equivalent to
what the default runner was before.

I think it unnecessary overhead to provide runner configuration to
customize runner behavior.  Anytime you wish to customize the runner,
you have to be sure to do everything the default runner was doing and
then keep that in sync.

Intead, the runner should provide hooks to add custom behavior.  This
gets the benefit of just having to maintain what is custom and not
what should be default.

In a coming PR, I'll be adding handler callbacks that only the default
runner will call.  This will be the place to customize runner behavior
on a per-handler basis.  This is setup for that work.