Feature branch, modify Sanford configuration and startup logic
This is an effort to modify Sanford to work more like Deas in how
it is configured and is started. The goal is to make them similar
and also simplify some of the oddities of working with Sanford.
This will change the CLI to take a configuration file, which is
expected to define a Sanford server. The server will contain all
the configuration for the IP, port and other options. The server
will have (or be given) a router. The router will be where
services are routed. Service handlers will continue to work as
they previously did. This branch is a collection of the changes to
get the changes working as expected.