Update to use the new DatTCP, rename server to server old
This updates Sanford to work with the latest DatTCP but also
renames server to server old. This is done to make room for
reworking the server without breaking sanfords full functionality.
Until we are ready to switch to the new server logic, the old
logic can continue to be function. This switches to using the
new DatTCP with leaving the server logic as unchanged as possible.
This doesn't change any functionality, just switches out to the
new DatTCP and renames the server.