Add new `Server` DSL
This adds a new `Server` and the DSL for configuring it. This is
part of the effort to rework the `Server` and startup process for
Sanford. This moves most of the `Host` configuration over to the
`Server` and the DSL for configuring it as well. This also moves
the template source configuration to the `Server` from the
global Sanford config. Both the `Host` and `Config` will be
removed later in favor of the new `Server`.