Build DatTCP server when initializing a Sanford server
This updates the new server logic to build a DatTCP server. This
allows the Sanford server to actually listen and serve connections
made to it. This adds the same socket handling behavior as the
old server logic, only it builds and runs a new worker. This also
adds all the same configuration and socket handling from the old
server logic. The TCP server will have the `TCP_NODELAY` option
set on it and in a linux environment, the client socket will
be "corked" as before. This finishes the main work on the
new `Server` logic and it can now replace the old server logic
in the test suite.

This also fixes a few mistakes made in previous commits. The
configuration now adds its routes and error procs to its hash.
Otherwise the config data won't ever get them. I also updated the
`router` DSL method to take a block so services could be added
to the router via the server definition.