Add `PIDFile` for managing a Sanford process PID file
This adds a helper class for managing a Sanford process PID file.
When a Sanford process is started, it writes its PID file to a
configured path. This is used to control the Sanford process when
using the CLI. When a stop, halt or restart command is given, the
PID file will be read and a TERM, INT or USR2 signal will be sent
to the process. This logic will be used by both the building and
running a Sanford process set of logic and the sending signals
to a Sanford process set of logic.