Add `ConfigFile` for evaluating Sanford config files
This adds the logic for evaluating Sanford config files. This will
be used to "choose" which server you want to run via the Sanford
CLI. These files work in a similar manner to Rack's rackup files.
The file expects `run` to be called with a Sanford server. The
CLI will then use the server to either start a process or send
a signal to an existing process.