Rework `CLI` to use `ConfigFile, `Process` and `ProcessSignal`
This reworks the `CLI` to use all the new logic. It now expects
a config file path and builds a `ConfigFile` to evaluate it. The
server from the `ConfigFile` is then either passed to a `Process`
or `ProcessSignal` depending on the command. Sanford supports the
same commands it did previously: run, start, stop and restart.
This makes all of Sanford's old logic no longer needed and it
will be removed in a future commit.