Remove old `Config`, `Host` and `Manager` logic
This removes the old Sanford logic related to running a Sanford
server using a `Manager`. This includes the `Sanford` module
configuration and init methods, `Config`, `Host`, `HostData`,
`Hosts` and `Manager`. With the switch of the CLI from the
`Manager` to the `Process` and `ProcessSignal` none of this logic
is used anymore. This is the final step of moving Sanford from
having a services file and hosts to having a config file with a
server configured in it.