Remove `Runner.run` and `ServiceHandler.run`
This removes the `Runner.run` and the `ServiceHandler.run` class
methods. These were a neat feature but haven't really worked out
or been used. Since these haven't been used, this removes them so
they no longer have to be maintained. With planned changes to the
`Runner.new` args, this avoids having to update both of these
methods to using the new args.