Add server system test for a service that renders a template
This adds a system test for a service that renders a template.
The goal of this test is to ensure that a template source
configured on the server can be used to render a template for
a service handler. This also fixes a bug where the configuration
was not passing the template source to server data. This fixes
that and ensures that configuring a template source and using it
to render a template works as expected.