Stop using `Rake` for running benchmarks
This changes the benchmarks to stop using `Rake` and just use
the Sanford CLI and ruby scripts. This minimizes using Rake with
the gem and only requires it for gem building. I'd rather not rely
on extra dependencies that are not needed and for this case, Rake
isn't needed.

This removes the "run X requests against the Benchmark server"
functionality. This hasn't been used since Sanford has been stable
and isn't related to "benchmarking". It was more of a system test
to see that everything worked when running a server.