Rename `FakeConnection` to `FakeServerConnection` and simplify
This renames the test support `FakeConnection` to
`FakeServerConnection` and simplifies it. The goal of this is to
clarify the difference between a fake protocol connection, which
is part of `Sanford::Protocol` and a fake server connection which
is part of `Sanford`. This also simplifies its implementation. It
was previously used for other tests and had extra behavior built
into it. It is now used only for server and worker unit tests so
its behavior could be scaled back.